Split mobile/desktop header files
